Reorganize sysfs tests by source file
Split the sysfs tests so each source file has a matching test file and
the package layout is easier to navigate.

Add a few focused router and worker edge-case tests while keeping the
existing sysfs coverage intact.
EOF && git push
